dbs/requirements.txt
python-keycloak==3.7.0
python-dotenv==1.0.0
python==3.12.0